refactor(agents): remove the unreachable ClaudeAdapter turn path (GH-1786)
ClaudeAdapter was only ever registered ControlOnly, so
turn_execution_adapter() never returned it and its start_turn — a
drifted duplicate of claude.rs CLI arg construction — was unreachable.
Its side-channel methods were Unsupported stubs (steer,
respond_approval) or a no-op interrupt (child handle was always None,
since only executed turns register live adapters).

- Delete ClaudeAdapter plus its spawn/heartbeat machinery; move the
  shared stream-json line parsers into claude_stream_json.rs
- Move the 18 parser tests; drop the 10 adapter arg-construction tests
- Remove the claude adapter registration from builder.rs and the two
  ClaudeAdapter entries from harness-cli/clippy.toml
- Drop the two-file sync obligation from CLAUDE.md

Acceptance: no duplicated arg construction, no sync obligation, and
cargo test --package harness-agents passes on the reduced surface
(228 passed; workspace clippy -D warnings clean).
