+++ crates/harness-cli/clippy.toml
@@ -11,8 +11,6 @@ disallowed-methods = [
     { path = "harness_agents::claude::ClaudeCodeAgent::with_reasoning_budget", reason = "Keep configured reasoning in harness_agents::builder." },
     { path = "harness_agents::claude::ClaudeCodeAgent::with_stream_timeout", reason = "Keep configured timeouts in harness_agents::builder." },
     { path = "harness_agents::claude::ClaudeCodeAgent::with_provider_backpressure_gate", reason = "Use harness_agents::builder::claude_agent_from_config or registry_from_config." },
-    { path = "harness_agents::claude_adapter::ClaudeAdapter::new", reason = "Use harness_agents::builder::registry_from_config." },
-    { path = "harness_agents::claude_adapter::ClaudeAdapter::with_provider_backpressure_gate", reason = "Use harness_agents::builder::registry_from_config." },
     { path = "harness_agents::codex::CodexAgent::new", reason = "Use harness_agents::builder::codex_agent_from_config or registry_from_config; cmd/pr.rs has the only reviewed exception." },
     { path = "harness_agents::codex::CodexAgent::with_cloud", reason = "Use harness_agents::builder::codex_agent_from_config or registry_from_config." },
     { path = "harness_agents::codex::CodexAgent::with_stream_timeout", reason = "Keep configured timeouts in harness_agents::builder; cmd/pr.rs has the only reviewed exception." },
